Increase "Show messages" time range
I've seen several customers who use the sidecar/winlogbeat to send in Windows EventLog messages and thinking it did not work because the "Show all messages" button triggers a search in only the last 8 hours. Reason for the empty search result is, that old EventLog messages are read first, and those can easily be more than 8 hours old. By increasing the limit to 7 days, we should always hit messages. The possible performance impact of a search over more indices is a calculated risk here and if your setup can't survive this kind of search, you probably have much big issues to solve.