Handle changing configuration as an array
Different collectors may use different configurations, so the
configuration selector needs to support several selected items, while
letting users to select only one option. Anyway, it is also possible for
users to remove a single configuration (by clicking on an active
configuration), and we need to handle this case as well.